Didn't initialize the random wallet quite right.
First, the long_to_hex call was wrong - it put 192 Bytes of zeros before
the 64 byte random key. I fixed this to only use the randomly generated
part.

Second, on the advice of @cyc I am using random.SystemRandom() instead
of just random. The reason is right there in the python docs:

http://docs.python.org/2/library/random.html:
"The pseudo-random generators of this module should not be used for
security purposes. Use os.urandom() or SystemRandom if you require a
cryptographically secure pseudo-random number generator."

yolocrypto